requirements.txt
requests==2.25.0
colorama
datetime
python-dotenv
